+++ chopper/lib/src/response.dart
@@ -1,3 +1,5 @@
+import 'dart:typed_data';
+
 import 'package:http/http.dart' as http;
 import 'package:meta/meta.dart';
 
@@ -8,8 +10,8 @@ class Response<Body> {
 
   const Response(this.base, this.body);
 
-  Response replace<Body>({http.Response base, Body body}) =>
-      Response(base ?? this.base, body ?? this.body);
+  Response replace<BodyType>({http.Response base, BodyType body}) =>
+      Response<BodyType>(base ?? this.base, body ?? this.body);
 
   Response<Body> replaceWithNull<Body>({http.Response base, Body body}) =>
       Response(base ?? this.base, body);
@@ -17,4 +19,8 @@ class Response<Body> {
   int get statusCode => base.statusCode;
 
   bool get isSuccessful => statusCode >= 200 && statusCode < 300;
+
+  Map<String, String> get headers => base.headers;
+
+  Uint8List get bodyBytes => base.bodyBytes;
 }