+++ chopper/lib/src/utils.dart
@@ -1,4 +1,5 @@
 import 'package:chopper/chopper.dart';
+import 'package:logging/logging.dart';
 
 Request applyHeader(Request request, String name, String value) =>
     applyHeaders(request, {
@@ -10,3 +11,9 @@ Request applyHeaders(Request request, Map<String, String> headers) {
   h.addAll(headers);
   return request.replace(headers: h);
 }
+
+final chopperLogger = Logger('Chopper');
+
+const contentTypeKey = 'content-type';
+const jsonHeaders = "application/json";
+const formEncodedHeaders = "application/x-www-form-urlencoded";