+++ chopper/CHANGELOG.md
@@ -1,16 +1,20 @@
 # Changelog
 
-## 1.2.0
+## 2.0.0
 
 - Request is now containing baseUrl
 - Can call `Request.toHttpRequest()` direclty to get the `http.BaseRequest` will receive
 - If a full url is specified in the `path` (ex: @Get(path: 'https://...')), it won't be concaten with the baseUrl of the ChopperClient and the ChopperAPI
 - Add `CurlInterceptor` thanks @edwardaux
 - Add `HttpLoggingInterceptor`
+- Add `FactoryConverter` annotation `@FactoryConverter(request: convertRequest, response: convertResponse)`
 
 - ***BreakingChange***
-  - remove `formUrlEncodedAPI`, this is the default behavior of http package and if `jsonApi` is false
-  - Method.url renamed to url
+  - Method.url renamed to path
+  - `Converter.encode` and `Converter.decode` removed, implement `Converter.convertResponse` and Converter.convertRequest` instead
+  - `ChopperClient.jsonApi` deprecated, use a `JsonConverter` instead
+  - `ChopperClient.formUrlEncodedApi`, use `FormUrlEncodedConverter` instead
+  - remove `JsonEncoded` annotation
 
 ## 1.1.0